created ingest and representation_creation folders, and moved files around
