+++ bia-ingest/scripts/create_specimens_and_updated_eids.py
@@ -8,8 +8,8 @@ It should only be needed for studies that were ingested before the 20/09/2024, a
 import typer
 from bia_ingest.config import api_client
 from bia_ingest.persistence_strategy import ApiPersister
-from bia_ingest.conversion import specimen, experimental_imaging_dataset
-from bia_ingest.biostudies import load_submission
+from bia_ingest.ingest import specimen, experimental_imaging_dataset
+from bia_ingest.ingest.biostudies import load_submission
 from bia_ingest.cli_logging import IngestionResult
 
 api_persister = ApiPersister(api_client)
